+++ src/main/scala/example/example.scala
@@ -1,7 +1,0 @@
-package scalajs.example
-
-object Example {
-  def main(args: Array[String]): Unit = {
-    println("Hello from Scala.js starter!")
-  }
-}